Fix bug with testing if field is filterable
The use of defaultdict necessitates slightly different logic to what's
in the upstream JS version. Previously, if you attempted to filter a
field that had no filter functions defined, and then test if it was
filterable, the "can" method would report back as True, rather than
False (as it would have before the filtering call). This issue has now
been resolved.